feat:new component filament resource

+++ database/migrations/2025_08_04_211853_create_orders_table.php
@@ -13,7 +13,7 @@ return new class extends Migration
     {
         Schema::create('orders', function (Blueprint $table) {
             $table->id();
-            $table->decimal('total', 10, 2)->default(0);
+            $table->integer('total')->default(0);
             $table->timestamps();
         });
     }
